+++ cpanel/both/routers/event.js
@@ -1,12 +1,12 @@
-Router.route('sample/myCustomerReport', function () {
-    this.render('sample_myCustomerReport');
+Router.route('cpanel/eventReport', function () {
+    this.render('cpanel_eventReport');
 }, {
-    name: 'sample.myCustomerReport',
-    header: {title: 'my customer', sub: '', icon: 'file-text-o'},
-    title: "My Customer"
+    name: 'capnel.eventReport',
+    header: {title: 'Event Report', sub: '', icon: 'file-text-o'},
+    title: "Event Report"
 });
 
-Router.route('sample/myCustomerReportGen', function () {
+Router.route('cpanel/eventReportGen', function () {
     // Config layout
     this.layout('reportLayout', {
         // Page size: a4, a5, mini | Orientation: portrait, landscape
@@ -16,8 +16,9 @@ Router.route('sample/myCustomerReportGen', function () {
         }
     });
 
+    // Generate
     var q = this.params.query;
-    this.render('sample_myCustomerReportGen', {
+    this.render('cpanel_eventReportGen', {
         data: function () {
             return q;
         }