+++ cpanel/client/app/subscriptions/subscriptions.js
@@ -1,7 +1,10 @@
+/**
+ * Global subscriptions
+ */
 Meteor.subscribe('cpanel_user');
+Meteor.subscribe('cpanel_setting');
 Meteor.subscribe('cpanel_company');
 Meteor.subscribe('cpanel_branch');
-Meteor.subscribe('cpanel_currency');
 Meteor.subscribe('cpanel_exchange');
-Meteor.subscribe('cpanel_setting');
+Meteor.subscribe('cpanel_currency');
 Meteor.subscribe('events');