+++ cpanel/client/templates/layout/main/header.js
@@ -1,4 +1,7 @@
-Template.layoutHeader.helpers({
+/**
+ * Helper
+ */
+Template.headerLayout.helpers({
     navbar: function () {
         var module = Session.get('currentModule');
         var branch = Session.get('currentBranch');
@@ -15,8 +18,22 @@ Template.layoutHeader.helpers({
             return {show: false};
         }
 
-        var titleBranch = Cpanel.Collection.Branch.findOne(branch);
-        var title = branch + ' : ' + titleBranch.enShortName;
+        var getBranch = Cpanel.Collection.Branch.findOne({_id: branch});
+        var title = branch + ' : ' + getBranch.enShortName;
         return {show: true, title: title};
     }
 });
+
+/**
+ * Server clock
+ */
+Template.headerLayout.onRendered(function () {
+    Meteor.setInterval(serverClock, 1000);
+});
+
+/* Clock function */
+function serverClock() {
+    Meteor.call('currentDate', function (error, result) {
+        $('#clock').html('<i class="fa fa-calendar"></i> ' + moment(result, 'YYYY-MM-DD H:mm:ss').format('ddd D, MMM YYYY H:mm:ss'));
+    });
+}