+++ cpanel/client/templates/layout/main/main.js
@@ -1,7 +1,7 @@
 /**
  * Layout
  */
-Template.layout.helpers({
+Template.mainLayout.helpers({
     appName: function () {
         var module = Session.get('currentModule');
         var branch = Session.get('currentBranch');
@@ -31,7 +31,7 @@ Template.layout.helpers({
     }
 });
 
-Template.layout.events({
+Template.mainLayout.events({
     'click .navbar-brand': function () {
         var module = Session.get('currentModule');
         if (!_.isUndefined(module)) {
@@ -43,18 +43,3 @@ Template.layout.events({
         }
     }
 });
-
-/**
- * Layout header
- */
-Template.layoutHeader.onRendered(function () {
-    Meteor.setInterval(serverClock, 1000);
-});
-
-/* Clock function */
-function serverClock() {
-    Meteor.call('currentDate', function (error, result) {
-        $('#clock').html('<i class="fa fa-calendar"></i> ' + moment(result, 'YYYY-MM-DD H:mm:ss').format('ddd D, MMM YYYY H:mm:ss'));
-    });
-}
-