+++ cpanel/server/startup/currency.js
@@ -1,12 +1,12 @@
 Meteor.startup(function () {
     if (Cpanel.Collection.Currency.find().count() == 0) {
         var doc = [
-            {_id: 'KHR', name: 'Cambodian Riel', symbol: '?'},
+            {_id: 'KHR', name: 'Cambodian Riel', symbol: 'R'},
             {_id: 'USD', name: 'United States Dollar', symbol: '$'},
             {_id: 'THB', name: 'Thai Baht', symbol: 'B'}
         ];
 
-        doc.forEach(function (obj) {
+        _.each(doc, function (obj) {
             Cpanel.Collection.Currency.insert(obj);
         });
     }