+++ sample/client/app/subscriptions/subscriptions.js
@@ -1,2 +1,5 @@
+/**
+ * Global subscriptions
+ */
 Meteor.subscribe('sample_customer');
 Meteor.subscribe('sample_address');