+++ sample/client/templates/reports/myCustomer/myCustomer.js
@@ -1,4 +1,0 @@
-Template.sample_myCustomerReport.onRendered(function () {
-    var $name = $('[name="date"]');
-    DateTimePicker.dateRange(name);
-});