+++ sample/server/collectionHooks/customer.js
@@ -1,19 +1,19 @@
 /**
  * Customer
  */
-Sample.Collection.Customer.before.insert(function (userId, doc) {
+Sample.Collection.Customer.after.insert(function (userId, doc) {
     Events.trackInsert({
         description: doc
     });
 });
 
-Sample.Collection.Customer.before.update(function (userId, doc, fieldNames, modifier, options) {
+Sample.Collection.Customer.after.update(function (userId, doc, fieldNames, modifier, options) {
     Events.trackUpdate({
-        description: modifier
+        description: EJSON.stringify(doc) + ' To ' + EJSON.stringify(modifier)
     });
 });
 
-Sample.Collection.Customer.before.remove(function (userId, doc) {
+Sample.Collection.Customer.after.remove(function (userId, doc) {
     Events.trackRemove({
         description: doc
     });